finalized the project
Moved the article loop into articles.php file so it can be used in the index.php and author.php. Added pictures to all the articles in data.php. changed the date format in data.php. gave all the authors a unique background color to their articles. also removed the individual background color in all the articles.  Added a footer.php and a header.php lastly cleaned up the main.css file